% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are.lmoms.valid.R
\name{are.lmoms.valid}
\alias{are.lmoms.valid}
\title{Validate L-moment Constraints}
\usage{
are.lmoms.valid(lmom, clean = TRUE, return_numeric = FALSE)
}
\arguments{
\item{lmom}{vector of L-moments}

\item{clean}{logical if it is \code{TRUE} (default) clean \code{lmom}'s names}

\item{return_numeric}{logical if it is \code{TRUE} L-moments corrected are returned instead of a boolean value.}
}
\value{
logical value indicating if all of these conditions are verified:
\itemize{
  \item \eqn{\lambda_2 > 0} (L-scale must be positive)
  \item \eqn{|\tau_3| < 1} (L-skewness must lie within \eqn{(-1, 1)})
  \item \eqn{|\tau_i| < 1} (i-th L-moment ratio must lie within \eqn{(-1, 1)} for \code{i>3})
  \item \eqn{\lambda_1} and \eqn{\lambda_2} must be finite
  \item \eqn{\frac{5\tau_3^2 - 1 }{4} \leq \tau_4 < 1} is checked for theoretical consistency
}
}
\description{
This function checks whether a given set of L-moments satisfies theoretical and practical constraints
commonly used in statistical modeling. These constraints ensure the validity of L-moment ratios and
help detect potential issues in the data or its transformation (e.g., log-transformed data with non-positive values).
}
\examples{

library(lmomPi)

data(airquality)
lmom <- samlmu(airquality$Ozone, 6)
are.lmoms.valid(lmom = lmom)

are.lmoms.valid(lmom = lmom,return_numeric=TRUE)

lmoml <- samlmu(airquality$Ozone, 6,ratio=FALSE)

are.lmoms.valid(lmom = lmoml)

are.lmoms.valid(lmom = lmoml,return_numeric=TRUE)


}
\references{
Hosking, J.R.M. and Wallis, J.R. (1997). \emph{Regional Frequency Analysis: An Approach Based on L-Moments}. Cambridge University Press. \doi{10.1017/CBO9780511529443}
}
